refactor: consolidate error handling imports across Yandex Cloud nodes
