+++ nodes/YandexCloudSpeechKit/YandexCloudSpeechKit.node.ts
@@ -9,8 +9,7 @@ import { NodeOperationError } from 'n8n-workflow';
 import { Session } from '@yandex-cloud/nodejs-sdk';
 import { ttsService, tts } from '@yandex-cloud/nodejs-sdk/dist/clients/ai-tts-v3/index';
 import { mapKeys, camelCase } from 'lodash';
-import { YandexCloudSdkError } from '@utils/sdkErrorHandling';
-import { withSdkErrorHandling } from '@utils/errorHandling';
+import { YandexCloudSdkError, withSdkErrorHandling } from '@utils/sdkErrorHandling';
 
 interface IIAmCredentials {
 	serviceAccountId: string;
